Move IIDs to `internal` library and some refactoring
