+++ packages/windows_foundation/lib/internal.dart
@@ -11,6 +11,7 @@ export 'src/iasyncoperation.dart';
 export 'src/internal/async_helpers.dart';
 export 'src/internal/comobject_pointer.dart';
 export 'src/internal/hstring_array.dart';
+export 'src/internal/iids.dart';
 export 'src/internal/int_array.dart';
 export 'src/internal/ipropertyvalue_helpers.dart';
 export 'src/internal/map_helpers.dart';
