+++ packages/windows_foundation/lib/src/collections/imap.dart
@@ -10,12 +10,9 @@ import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
 import 'package:windows_media/windows_media.dart';
 
-import '../constants.dart';
+import '../../internal.dart';
 import '../helpers.dart';
 import '../iinspectable.dart';
-import '../internal/ipropertyvalue_helpers.dart';
-import '../internal/map_helpers.dart';
-import '../ipropertyvalue.dart';
 import '../types.dart';
 import '../winrt_enum.dart';
 import 'iiterable.dart';
