--- packages/windows_foundation/lib/src/constants.dart
@@ -1,1000 +1,0 @@
-// Copyright (c) 2023, the dartwinrt authors. Please see the AUTHORS file for
-// details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-// Constants used by WinRT projection
-
-// ignore_for_file: constant_identifier_names
-
-// IIterable<IKeyValuePair<K, V>> IIDs
-const IID_IIterable_IKeyValuePair_Guid_BackgroundTaskRegistration =
-    '{62ae0fda-b238-554f-a275-1dc16d6ca03a}';
-const IID_IIterable_IKeyValuePair_Guid_IBackgroundTaskRegistration =
-    '{80fb0327-5a00-55cc-85db-a852719981b9}';
-const IID_IIterable_IKeyValuePair_Guid_Object =
-    '{f3b20528-e3b3-5331-b2d0-0c2623aee785}';
-const IID_IIterable_IKeyValuePair_Guid_SpatialSurfaceInfo =
-    '{868757d1-be21-51d9-8dee-a958b9deec71}';
-const IID_IIterable_IKeyValuePair_Object_Object =
-    '{33ac68f0-1084-529a-8a17-4e7c8adb7a0c}';
-const IID_IIterable_IKeyValuePair_PedometerStepKind_PedometerReading =
-    '{098f29cb-bc91-5639-a541-d5a4811ec35b}';
-const IID_IIterable_IKeyValuePair_String_AppCapabilityAccessStatus =
-    '{62e88ad9-d63e-5173-baa2-bb4521c7e82a}';
-const IID_IIterable_IKeyValuePair_String_ApplicationDataContainer =
-    '{a785be1d-159e-53ad-9553-598b03dca048}';
-const IID_IIterable_IKeyValuePair_String_BackgroundTaskRegistrationGroup =
-    '{04428524-7d54-59b4-bf17-ac57c4ce6b40}';
-const IID_IIterable_IKeyValuePair_String_BitmapTypedValue =
-    '{05f9430c-2f22-5638-aa89-8c9abcd54ff9}';
-const IID_IIterable_IKeyValuePair_String_ChatMessageStatus =
-    '{57d87c13-48e9-546f-9b4e-a3906e1e7c24}';
-const IID_IIterable_IKeyValuePair_String_GattServiceProviderConnection =
-    '{eb2cecfb-dd6c-5329-8e17-7fb80a84191f}';
-const IID_IIterable_IKeyValuePair_String_IBuffer =
-    '{3c9ffa92-5123-5ac4-b111-03c215f0c51c}';
-const IID_IIterable_IKeyValuePair_String_ICompositionAnimationBase =
-    '{c739c420-64b2-53f2-89b9-05906aba7cb0}';
-const IID_IIterable_IKeyValuePair_String_IJsonValue =
-    '{dfabb6e1-0411-5a8f-aa87-354e7110f099}';
-const IID_IIterable_IKeyValuePair_String_IPrintOptionDetails =
-    '{6770cf39-094f-59c5-8a5d-e3b5dc64db0f}';
-const IID_IIterable_IKeyValuePair_String_IVectorView_String =
-    '{a4cd6151-2cc1-56f1-9014-df6ba3410beb}';
-const IID_IIterable_IKeyValuePair_String_IVectorView_TextSegment =
-    '{f819a276-b3f5-54d4-b8fd-c9adb7f700e3}';
-const IID_IIterable_IKeyValuePair_String_IVector_String =
-    '{4fed2669-d0d3-59f6-91d9-95902d728d6a}';
-const IID_IIterable_IKeyValuePair_String_IppAttributeError =
-    '{17f42ebf-08e9-5e26-96ed-513171d2aff2}';
-const IID_IIterable_IKeyValuePair_String_IppAttributeValue =
-    '{5dcf9ab4-ed4e-5648-8bfb-f626d5d7c505}';
-const IID_IIterable_IKeyValuePair_String_LicenseSatisfactionInfo =
-    '{ac935021-e04b-5226-8119-5b73d0b8be5b}';
-const IID_IIterable_IKeyValuePair_String_MediaFrameSource =
-    '{a038e80f-0b3d-5cd4-849d-13036e1f506e}';
-const IID_IIterable_IKeyValuePair_String_NamedResource =
-    '{bf16482e-80ed-51f0-b9c9-3a804e2d6403}';
-const IID_IIterable_IKeyValuePair_String_Object =
-    '{fe2f3d47-5d47-5499-8374-430c7cda0204}';
-const IID_IIterable_IKeyValuePair_String_ProductLicense =
-    '{ffe05002-6e65-5001-a752-9b27ed3e2839}';
-const IID_IIterable_IKeyValuePair_String_ProductListing =
-    '{3a7d9a2f-2a82-59f8-bd7c-d691ca169863}';
-const IID_IIterable_IKeyValuePair_String_RandomAccessStreamReference =
-    '{c9729ba7-5e20-569d-a3d1-97a4e653e5bb}';
-const IID_IIterable_IKeyValuePair_String_ResourceMap =
-    '{9895431e-a817-5509-91e6-d90a933ed830}';
-const IID_IIterable_IKeyValuePair_String_SceneAttributeSemantic =
-    '{a3e30221-7ca2-5a3c-a54a-378fee7369cc}';
-const IID_IIterable_IKeyValuePair_String_SocketActivityInformation =
-    '{3e43fa16-7af1-51df-a0d3-da81b321639d}';
-const IID_IIterable_IKeyValuePair_String_SpatialAnchor =
-    '{55f0fa8a-afd4-5541-a1c3-36f12147d606}';
-const IID_IIterable_IKeyValuePair_String_StoreLicense =
-    '{ca8ba445-6f4d-5da9-95ee-42cf118def63}';
-const IID_IIterable_IKeyValuePair_String_StoreProduct =
-    '{78a33722-abfb-57c0-853f-5616a3ab8d57}';
-const IID_IIterable_IKeyValuePair_String_String =
-    '{e9bdaaf0-cbf6-5c72-be90-29cbf3a1319b}';
-const IID_IIterable_IKeyValuePair_String_TargetedContentValue =
-    '{45a020d8-fe49-5720-950b-3cceab655531}';
-const IID_IIterable_IKeyValuePair_String_VoiceCommandDefinition =
-    '{67693dd1-ac45-5ef3-9ba6-4d78709d9ee0}';
-const IID_IIterable_IKeyValuePair_String_VoiceCommandSet =
-    '{86e4878a-45de-5203-be77-81c04c9b44e6}';
-const IID_IIterable_IKeyValuePair_String_WalletItemCustomProperty =
-    '{3f45154c-0c0d-5df9-a557-259f20c927ae}';
-const IID_IIterable_IKeyValuePair_String_WalletRelevantLocation =
-    '{e55e510c-4028-5df0-b78b-27bd06980b0b}';
-const IID_IIterable_IKeyValuePair_String_WalletTransaction =
-    '{4fd2ca58-b7a2-5923-9380-49c11c69c39a}';
-const IID_IIterable_IKeyValuePair_String_WalletVerb =
-    '{6d2ef172-8ae3-572e-ae44-8ee3fd49d19f}';
-const IID_IIterable_IKeyValuePair_UINT32_IBuffer =
-    '{4fe7fe23-22b1-528c-881d-a4eceaef0f11}';
-
-// IIterable<T> IIDs
-const IID_IIterable = '{faa585ea-6214-4217-afda-7f46de5869b3}';
-const IID_IIterable_AccessListEntry = '{4995c2b0-736b-588d-ae42-6f69b025b388}';
-const IID_IIterable_ActivationSignalDetectionConfiguration =
-    '{cb03ac8c-3a88-5c35-9872-00a466e49a12}';
-const IID_IIterable_ActivationSignalDetectionTrainingDataFormat =
-    '{88b67e2c-ec3f-520c-ae65-bfe63272fadd}';
-const IID_IIterable_ActivationSignalDetector =
-    '{e1f74a8f-38eb-58be-b770-78af5a4a6b2e}';
-const IID_IIterable_ActivationSignalDetectorPowerState =
-    '{b6860d48-cec5-5a00-8653-4c924a991575}';
-const IID_IIterable_ActivitySensorReading =
-    '{9a34ce03-8c6d-5994-907f-d5c2d19148cb}';
-const IID_IIterable_ActivitySensorReadingChangeReport =
-    '{551a4962-9e96-5e6b-8b8a-65ee3d0046f3}';
-const IID_IIterable_ActivityType = '{2a04cdfa-5dfd-5178-8731-ade998e4a7f6}';
-const IID_IIterable_AdcController = '{4e478aad-4861-5758-b64b-5b4f28d8f86e}';
-const IID_IIterable_AdvancedPhotoMode =
-    '{7d090784-70a9-570c-be82-0d0890318975}';
-const IID_IIterable_AllJoynAuthenticationMechanism =
-    '{d307c7af-4106-5d1c-b06c-5eb593d9be34}';
-const IID_IIterable_AlternateWordForm =
-    '{ae838c29-60c1-5093-ae61-0ed22857db05}';
-const IID_IIterable_AppBroadcastPlugIn =
-    '{c531c5b0-0223-5c9e-a8fb-20f52ea58d75}';
-const IID_IIterable_AppDiagnosticInfo =
-    '{8118de8f-3ae3-55e1-80a8-25453cdba894}';
-const IID_IIterable_AppExtension = '{3b4fe356-1b13-59cb-ab1f-c4667a74756b}';
-const IID_IIterable_AppInfo = '{63d0bffe-0e34-55b3-83d5-314caff2b137}';
-const IID_IIterable_AppInstallItem = '{93e1cdc8-503f-55b1-915b-c0dc7888ce31}';
-const IID_IIterable_AppInstance = '{2c28b4c9-6378-5382-a1a6-79bd6eed967a}';
-const IID_IIterable_AppListEntry = '{86f4d4ef-d8fd-5fb5-807c-72da8fc9e544}';
-const IID_IIterable_AppRecordingSavedScreenshotInfo =
-    '{dd170424-794d-5158-a9af-6824353f91b2}';
-const IID_IIterable_AppResourceGroupBackgroundTaskReport =
-    '{3e7dcbca-1804-5672-ad3b-58d944bb044c}';
-const IID_IIterable_AppResourceGroupInfo =
-    '{8b640948-c0d3-5b7e-a99c-5956190d5408}';
-const IID_IIterable_AppUriHandlerHost =
-    '{f6e575bd-eab7-5981-ba01-49ff976ca186}';
-const IID_IIterable_Appointment = '{b9802bba-ff53-5d37-8cd7-e56162f12156}';
-const IID_IIterable_AppointmentCalendar =
-    '{d3fb010b-b692-5130-9d16-2cfdabcb6dec}';
-const IID_IIterable_AppointmentException =
-    '{caac41d6-6c65-5fd8-b783-eb9d9a4272b8}';
-const IID_IIterable_AppointmentInvitee =
-    '{fb9718a9-f059-52b0-a904-1a65e4281e40}';
-const IID_IIterable_AppointmentStoreChange =
-    '{7b657ca2-e02e-5026-a032-9905e49682fd}';
-const IID_IIterable_ArcadeStick = '{9376f457-2da5-544a-a409-c636f5d81c35}';
-const IID_IIterable_AttributedNetworkUsage =
-    '{d061dcb9-6854-5ef9-8e03-008a7a704c48}';
-const IID_IIterable_AudioDeviceModule =
-    '{7eeb51c3-d70e-548a-85c2-3cf71b4a124c}';
-const IID_IIterable_AudioEffect = '{07af9afd-25b8-579d-be7e-8acc03418d0b}';
-const IID_IIterable_AudioEncodingProperties =
-    '{00939468-25d8-533f-854e-4f20f36c51dc}';
-const IID_IIterable_AudioGraphConnection =
-    '{96168d06-a51a-5480-9403-fbd7631e3b3c}';
-const IID_IIterable_AudioStreamDescriptor =
-    '{a3e2c972-a171-5b94-8389-e983ebc3f3b9}';
-const IID_IIterable_AudioTrack = '{b3be50a8-9856-5656-ab6d-cd58f9de0a4f}';
-const IID_IIterable_AutoFocusRange = '{751664c6-f8d6-50a3-ab80-137c6d908c55}';
-const IID_IIterable_BackgroundAudioTrack =
-    '{3fc05ba2-30f6-5219-9047-1197ffae8dba}';
-const IID_IIterable_BackgroundTransferContentPart =
-    '{cf303199-de3b-5dac-a703-6c57d80821c4}';
-const IID_IIterable_BackgroundTransferFileRange =
-    '{2cc2d499-974c-5078-89ae-2d4ee1139721}';
-const IID_IIterable_BasicGeoposition = '{922399a8-0093-5009-a8d2-f87b0eae75f5}';
-const IID_IIterable_BitmapCodecInformation =
-    '{2b6bdb90-a4eb-5142-b582-3ccb1edc5789}';
-const IID_IIterable_BitmapPixelFormat =
-    '{e924d9ed-a13e-5bdb-9ed8-65a1474dc274}';
-const IID_IIterable_BluetoothLEAdvertisementBytePattern =
-    '{1e3fadee-54ac-538b-8777-351afb78cb74}';
-const IID_IIterable_BluetoothLEAdvertisementDataSection =
-    '{db98b5d1-897e-59cc-b86a-7b8855ac98af}';
-const IID_IIterable_BluetoothLEAdvertisementReceivedEventArgs =
-    '{34f6412f-8314-5205-967c-db357c9a42a7}';
-const IID_IIterable_BluetoothLEManufacturerData =
-    '{834a4cac-bb8b-5f0f-9f28-4dbc98c17907}';
-const IID_IIterable_CaptureSceneMode = '{16d26b98-2cbc-52f0-ab64-1723714418e9}';
-const IID_IIterable_CastingSource = '{1abb2cc9-46a2-58b1-91aa-28699d66d1ab}';
-const IID_IIterable_Certificate = '{0c7d1423-e8fd-5a91-b55c-8bfbe7ac2d40}';
-const IID_IIterable_CertificateExtension =
-    '{1bdd7127-73b3-5192-8bde-20c136281260}';
-const IID_IIterable_ChainValidationResult =
-    '{2628f58f-3f02-54f2-808f-e1117709d6d0}';
-const IID_IIterable_CharacterGrouping =
-    '{82e3abf0-06e3-5609-ba39-c51eb2f5fae6}';
-const IID_IIterable_ChatConversation = '{af43c676-a512-5388-9f69-0308953d719b}';
-const IID_IIterable_ChatMessage = '{647bf12c-f621-5fd8-af39-c72b16baf07e}';
-const IID_IIterable_ChatMessageAttachment =
-    '{8a1e36ac-13b6-577c-9b43-3d24a453f866}';
-const IID_IIterable_ChatMessageChange =
-    '{0df896df-354c-5d35-b7ac-6b2e1d70c8eb}';
-const IID_IIterable_ChatMessageTransport =
-    '{aae32f12-ed85-528c-8bad-6362d876ef8b}';
-const IID_IIterable_ChatRecipientDeliveryInfo =
-    '{e3d7b44c-4829-561b-a15e-a745e5adf6d1}';
-const IID_IIterable_ClipboardHistoryItem =
-    '{5c3705f4-8e46-5ae5-97bf-88220cccffd9}';
-const IID_IIterable_CmsSignerInfo = '{6af24174-2dda-5a54-a0b9-4d6690059427}';
-const IID_IIterable_CodecInfo = '{552f7709-3bc3-59fb-93e6-1ebd28d3c008}';
-const IID_IIterable_ConnectionProfile =
-    '{34dabef9-87d0-5b1c-a7ac-9d290adeb0c8}';
-const IID_IIterable_ConnectivityInterval =
-    '{58051a8b-b259-5414-9b9a-caa0789e833e}';
-const IID_IIterable_Contact = '{63319996-7e0f-552e-872b-7b9adb1f4997}';
-const IID_IIterable_ContactAddress = '{e783287c-e920-5540-9182-c1a6a5f2ad36}';
-const IID_IIterable_ContactAnnotation =
-    '{ea7fc05e-fe09-5690-a70b-a14867040fa5}';
-const IID_IIterable_ContactAnnotationList =
-    '{ae095a89-8b8b-5b8a-8f28-555aa490ad87}';
-const IID_IIterable_ContactChange = '{f761174c-1f89-5796-9d09-7fee0813704d}';
-const IID_IIterable_ContactConnectedServiceAccount =
-    '{35e203b2-5138-5967-80fa-a477ed4c7ddc}';
-const IID_IIterable_ContactDate = '{96045d0b-b38b-5d91-8311-4fbc7ad4337d}';
-const IID_IIterable_ContactEmail = '{37d0d55f-e136-5b07-8a3a-8110bde33917}';
-const IID_IIterable_ContactField = '{3b012111-c82b-541e-a0c1-37713ed83541}';
-const IID_IIterable_ContactFieldType = '{384b8b1b-ce8e-5781-b3dc-0776d684f658}';
-const IID_IIterable_ContactGroup = '{9db1688a-f7e0-5059-989e-d1453056d73a}';
-const IID_IIterable_ContactInformation =
-    '{352762af-17e3-55e0-8be1-6a598a44b5b4}';
-const IID_IIterable_ContactInstantMessageField =
-    '{810533b2-28bb-5ed3-ac19-fec8059d7b3c}';
-const IID_IIterable_ContactJobInfo = '{17c1cead-56a3-5d93-abf2-14d34d4b9356}';
-const IID_IIterable_ContactList = '{f9b0782c-bfe9-564e-bca2-97235fd64463}';
-const IID_IIterable_ContactLocationField =
-    '{2a0a41a6-da96-50a5-b8df-e4a8c4fcf86b}';
-const IID_IIterable_ContactMatchReason =
-    '{0a8c1bd9-2402-5c98-be6e-63a4808047ba}';
-const IID_IIterable_ContactPhone = '{4abd089c-49a1-5ba9-80df-53b72dd691ce}';
-const IID_IIterable_ContactSignificantOther =
-    '{ae12bd35-93d6-53d8-87c0-49d84c5c6a83}';
-const IID_IIterable_ContactWebsite = '{4b03c203-989e-5b97-8fcb-67d661defe76}';
-const IID_IIterable_ConversationalAgentVoiceActivationPrerequisiteKind =
-    '{4e9bb4c0-65ab-5049-9336-918b6d359071}';
-const IID_IIterable_CoreApplicationView =
-    '{32bc12d1-2653-5a41-a55e-88a12af2026a}';
-const IID_IIterable_CortanaPermission =
-    '{36a12eae-2e24-5e07-bfd0-344a92990916}';
-const IID_IIterable_DateTime = '{576a207d-977c-5b36-b54d-624ec86c53a3}';
-const IID_IIterable_DetectedFace = '{25347323-3556-5cbb-9855-2b5856437f4d}';
-const IID_IIterable_DeviceAccountConfiguration =
-    '{a9c3ea6e-9dd9-52fe-9d27-f9e4dedd4d3f}';
-const IID_IIterable_DeviceClass = '{47d4be05-58f1-522e-81c6-975eb4131bb9}';
-const IID_IIterable_DeviceInformation =
-    '{dd9f8a5d-ec98-5f4b-a3ea-9c8b5ad53c4b}';
-const IID_IIterable_DeviceWatcherEvent =
-    '{b48fd051-eafa-523f-a66e-9d4151c5d522}';
-const IID_IIterable_DeviceWatcherEventKind =
-    '{f04365ab-d3f3-5f85-a7da-dc19cff73d86}';
-const IID_IIterable_DigitalWindowCapability =
-    '{a4f28d49-fb50-5c0d-ba48-6695e4612500}';
-const IID_IIterable_DirectXPixelFormat =
-    '{3908f2c6-1aee-5129-b9a6-2a6e01d9507e}';
-const IID_IIterable_DisplayAdapter = '{db889bc5-3fd5-5b02-a2ea-d8daab6c33f2}';
-const IID_IIterable_DisplayModeInfo = '{dabee849-f1c8-5dab-abe5-effa5110894c}';
-const IID_IIterable_DisplayPath = '{4c0352b3-bb9e-52a7-afa7-c946928c12c9}';
-const IID_IIterable_DisplayTarget = '{ccb2e527-3a37-50fd-8d28-59e9db8cf9d9}';
-const IID_IIterable_DisplayView = '{8b18679f-3a2d-5d78-8409-2af682e7433c}';
-const IID_IIterable_DnssdServiceInstance =
-    '{1de3a3e5-387e-5328-b864-3f0e3475d343}';
-const IID_IIterable_DownloadOperation =
-    '{f6fd69cb-e6e7-56d5-9be6-e0dc4683fa80}';
-const IID_IIterable_ESimDiscoverEvent =
-    '{83504c13-a417-5601-9adb-f1ff18294dc9}';
-const IID_IIterable_ESimProfile = '{2ce4dd28-5154-57da-8e92-c1c9c964427b}';
-const IID_IIterable_EmailAttachment = '{a5ab00e4-0ed7-52ee-aacb-83d840d6f8fa}';
-const IID_IIterable_EmailCertificateValidationStatus =
-    '{7e326530-7449-51a7-b1bc-c43533a78e06}';
-const IID_IIterable_EmailConversation =
-    '{d495f8dc-4327-5bca-a506-44b172d6d435}';
-const IID_IIterable_EmailFolder = '{6dcb53eb-8a35-54bf-8c8e-449bce9b3920}';
-const IID_IIterable_EmailMailbox = '{8fc76b2e-f800-5f1d-a023-de47e5f306e0}';
-const IID_IIterable_EmailMailboxAction =
-    '{26419311-1cd7-5bab-9c4a-e95884d5539c}';
-const IID_IIterable_EmailMailboxChange =
-    '{4630ccb6-dfce-57dc-a5d1-28b889c2c2b7}';
-const IID_IIterable_EmailMessage = '{a5fac823-8e6d-5bbe-ad16-8350aea446e0}';
-const IID_IIterable_EmailRecipient = '{5f18cab2-236d-5ec5-bc64-e3e63d29e774}';
-const IID_IIterable_EmailRecipientResolutionResult =
-    '{cae3c1c4-c689-5787-976f-1a158ffdd16b}';
-const IID_IIterable_EmbeddedAudioTrack =
-    '{32fabed0-e1e6-578e-86e5-b4e6abeb22d6}';
-const IID_IIterable_EndpointPair = '{d7ec83c4-a17b-51bf-8997-aa33b9102dc9}';
-const IID_IIterable_EnhancedWaypoint = '{d0545dba-9b05-5e37-bfc0-3da2b51d135b}';
-const IID_IIterable_EnterpriseKeyCredentialRegistrationInfo =
-    '{e7eea796-77f9-5473-a913-734ea0e3ff46}';
-const IID_IIterable_EqualizerBand = '{6f76d148-023e-565a-9f09-4ad4a32ad74f}';
-const IID_IIterable_FileInformation = '{dd96d7e9-892b-5932-b677-5bc32588008f}';
-const IID_IIterable_FlightStick = '{3b7fc175-bebe-52ef-a3e9-dda75ea1acfc}';
-const IID_IIterable_FocusMode = '{561bc21f-4ae2-580a-a216-0ad48f373a4c}';
-const IID_IIterable_FocusPreset = '{26ba711b-3a32-5216-bc34-61ecafbebdc1}';
-const IID_IIterable_FolderInformation =
-    '{5c720bf5-7636-51fd-9ef7-d5f57f071a9b}';
-const IID_IIterable_ForceFeedbackMotor =
-    '{c14440d1-fea0-5147-aed8-9b85239da882}';
-const IID_IIterable_FrameController = '{bd8eeadc-2dd9-5ad8-ac5d-f3b13b94b9c2}';
-const IID_IIterable_Guid = '{f4ca3045-5dd7-54be-982e-d88d8ca0876e}';
-const IID_IIterable_GameListEntry = '{42b8c8a0-3d03-5d5f-817e-4405c850f646}';
-const IID_IIterable_GameSaveBlobInfo = '{a7c456d7-fa9f-536f-8ed2-459545811ed4}';
-const IID_IIterable_GameSaveContainerInfo =
-    '{55e4d98f-0889-5c06-a857-7dd168c2d852}';
-const IID_IIterable_Gamepad = '{47132ba0-6b17-5cd2-a8bd-b5d3443ccb13}';
-const IID_IIterable_GattCharacteristic =
-    '{e3c56728-7f2d-5a0d-ad38-030d39c60f9f}';
-const IID_IIterable_GattClientNotificationResult =
-    '{d3f4b8ad-251f-5bbb-ada2-ea47434e74d6}';
-const IID_IIterable_GattDescriptor = '{30e2ffc4-3aa3-5219-9a18-ca2d0b6562e8}';
-const IID_IIterable_GattDeviceService =
-    '{4b192e23-4893-56b2-8eff-439c3ab7fd1f}';
-const IID_IIterable_GattLocalCharacteristic =
-    '{7082ed53-37f6-5262-8d09-939bea9edbd5}';
-const IID_IIterable_GattLocalDescriptor =
-    '{9016e9a2-c9f7-5d09-b0ae-552fa67796ae}';
-const IID_IIterable_GattPresentationFormat =
-    '{d75fcef1-c10e-5b7b-b130-f5a00314d35d}';
-const IID_IIterable_GattSubscribedClient =
-    '{b19864e4-f2b1-50de-8c11-fff9eca934e9}';
-const IID_IIterable_GattValueChangedEventArgs =
-    '{526a63df-8827-51b9-9e2c-9d65021a79d7}';
-const IID_IIterable_GazePointPreview = '{1cf68266-3eb7-5336-840a-3c1d9fdf5349}';
-const IID_IIterable_Geofence = '{ce697733-595c-51c0-ad5f-324af5cdf2dd}';
-const IID_IIterable_GeofenceStateChangeReport =
-    '{76f50b4e-7aa7-565b-aada-b0c1cc144ed0}';
-const IID_IIterable_Geopath = '{246fcfbd-a881-5e8e-9908-c1b9ebdfec78}';
-const IID_IIterable_Geopoint = '{e7617fc9-2cc7-5bd1-bc5a-f47260834ed8}';
-const IID_IIterable_Geoposition = '{135ed72d-75b1-5881-be41-6ffeaa202044}';
-const IID_IIterable_Geovisit = '{d5800189-0f3f-54a0-a749-6000c1e12e58}';
-const IID_IIterable_GlobalSystemMediaTransportControlsSession =
-    '{6c594bfe-b3dd-5f1d-a78f-3a2d9e937ca3}';
-const IID_IIterable_GpioChangeRecord = '{b4afbf4f-620e-5725-878a-78c6ed10374e}';
-const IID_IIterable_GpioController = '{415c3794-b2b6-5f5c-9a05-ae9268514726}';
-const IID_IIterable_GuidanceLaneInfo = '{45960d72-1bf6-5a1d-a17f-e83f56f1ab57}';
-const IID_IIterable_GuidanceManeuver = '{b5780d67-8a8b-558f-a4b6-c4531ef32ec8}';
-const IID_IIterable_GuidanceRoadSegment =
-    '{f7c614c4-0fca-5eda-804c-85c829956334}';
-const IID_IIterable_HdmiDisplayMode = '{497e3d51-0ea1-5be0-8dba-8f7f4ce4fb33}';
-const IID_IIterable_HdrVideoMode = '{1d9679a7-2d06-5294-ac67-f9cd3561dcb8}';
-const IID_IIterable_HidBooleanControl =
-    '{1111e585-5ab0-5d2b-8aed-b6d6186d1c3f}';
-const IID_IIterable_HidBooleanControlDescription =
-    '{d0ff0fed-a156-58bf-9411-5777df9d57bf}';
-const IID_IIterable_HidCollection = '{bbeada0f-708f-5b5e-a017-5c64ffb96b69}';
-const IID_IIterable_HidNumericControlDescription =
-    '{868f060d-e0d4-571b-b2f7-431d6984a513}';
-const IID_IIterable_HolographicCamera =
-    '{b2afd154-8db0-5bb2-ad7a-684afd479264}';
-const IID_IIterable_HolographicCameraPose =
-    '{92111aff-8dcc-538e-ae3d-31fd252a0ad5}';
-const IID_IIterable_HolographicDepthReprojectionMethod =
-    '{2803c19f-8082-52c0-aad1-68ab51ed287f}';
-const IID_IIterable_HolographicFramePresentationReport =
-    '{d0fe1cdf-33ad-5051-8c5b-ab9a2b2c24bf}';
-const IID_IIterable_HolographicFrameScanoutReport =
-    '{cfa9ea14-4803-5001-9eda-cb4ffcd270e9}';
-const IID_IIterable_HolographicQuadLayer =
-    '{84744661-94de-5866-a15d-9efb19a99a54}';
-const IID_IIterable_HostName = '{9e5f3ed0-cf1c-5d38-832c-acea6164bf5c}';
-const IID_IIterable_HttpChallengeHeaderValue =
-    '{17847e58-f0cf-52e7-81c0-9d06a081569f}';
-const IID_IIterable_HttpConnectionOptionHeaderValue =
-    '{1afc512f-dfa2-5528-8aef-6e6a37789b3c}';
-const IID_IIterable_HttpContentCodingHeaderValue =
-    '{0e8960c2-76e5-56c1-9fe6-8cd9018a6786}';
-const IID_IIterable_HttpContentCodingWithQualityHeaderValue =
-    '{916e6271-cef6-53b8-b93f-dc8dca559e50}';
-const IID_IIterable_HttpCookie = '{0eb9fa36-88de-590d-8ea0-b613d0ab015f}';
-const IID_IIterable_HttpCookiePairHeaderValue =
-    '{16773244-5011-5244-bb00-14db4cf71005}';
-const IID_IIterable_HttpDiagnosticSourceLocation =
-    '{4286ca1a-a4c5-5ae8-9da9-5bfa24768e22}';
-const IID_IIterable_HttpExpectationHeaderValue =
-    '{0a391c2b-13c3-59b8-8662-13affb9d1754}';
-const IID_IIterable_HttpLanguageRangeWithQualityHeaderValue =
-    '{0c5a4a0d-dbb7-5abf-a5fd-c3b98d9c1487}';
-const IID_IIterable_HttpMediaTypeWithQualityHeaderValue =
-    '{ffc7dfde-58b8-51e9-b33b-65a5d6a4e24c}';
-const IID_IIterable_HttpMethod = '{c088bb1a-dc61-57e1-b967-78dab3541da5}';
-const IID_IIterable_HttpNameValueHeaderValue =
-    '{5838a126-b617-5417-b32e-ded2d27d40a3}';
-const IID_IIterable_HttpProductInfoHeaderValue =
-    '{bf97cf99-e614-5765-bad7-161ae3a93ae9}';
-const IID_IIterable_HttpTransferCodingHeaderValue =
-    '{57402048-38bf-5055-acb6-5c4da765e388}';
-const IID_IIterable_I2cController = '{a5ee8233-2429-5b26-9a02-993e4e7edfa9}';
-const IID_IIterable_IAdcControllerProvider =
-    '{30047155-1f71-5223-8482-e5159d0137d0}';
-const IID_IIterable_IAudioEffectDefinition =
-    '{607a20bf-32b4-5b8e-a793-3024f8d3582a}';
-const IID_IIterable_IBasicDevice = '{7d468b5e-763b-59cd-a086-ec6d8be0d858}';
-const IID_IIterable_IBuffer = '{902972bf-a984-5443-b1c5-2f04a99e1fca}';
-const IID_IIterable_IChatItem = '{ea8e3309-099a-592e-9e28-441b151fe061}';
-const IID_IIterable_IContactField = '{e065edc5-1d23-5be5-9998-2736cd4d5297}';
-const IID_IIterable_IDeviceIcon = '{16077ee6-dcfc-53aa-ab0e-d666ac819d6c}';
-const IID_IIterable_IDirect3DSurface = '{cc63bf9c-e16a-5a75-a5aa-2b53f975b0b0}';
-const IID_IIterable_IGpioControllerProvider =
-    '{09212bd4-851b-52bd-b82c-421bf3d6f511}';
-const IID_IIterable_IHttpContent = '{f102157f-b482-5736-9d12-c683bc494942}';
-const IID_IIterable_II2cControllerProvider =
-    '{11341a6c-3a02-5f73-9db8-c3ec5823e35d}';
-const IID_IIterable_IIndexableContent =
-    '{4a6edbfe-0c41-5042-ac58-a885a8fc7928}';
-const IID_IIterable_IIterable_IKeyValuePair_String_IppAttributeValue =
-    '{e4a5d51b-0047-5c20-8a42-d265732e8117}';
-const IID_IIterable_IIterable_Point = '{ae44597e-d411-5b7f-bbec-6a96c94a107a}';
-const IID_IIterable_IJsonValue = '{cb0492b6-4113-55cf-b2c5-99eb428ba493}';
-const IID_IIterable_ILampArrayEffect = '{3f108d37-6679-5590-aed2-033362fbf413}';
-const IID_IIterable_ILearningModelFeatureDescriptor =
-    '{0fa50877-6792-56b7-af46-430a8901894a}';
-const IID_IIterable_ILearningModelVariableDescriptorPreview =
-    '{df23db35-f789-51a1-856d-87cd7cd042f1}';
-const IID_IIterable_IMapView_String_IppAttributeValue =
-    '{47558612-ba48-5da9-9d10-6ed53b0ca405}';
-const IID_IIterable_IMapView_String_Object =
-    '{e1670fae-49cd-5c47-a8c8-f6fa2c650c6c}';
-const IID_IIterable_IMediaCue = '{737a2098-41bf-5d9f-a1d2-2f171c578b3c}';
-const IID_IIterable_IMediaEncodingProperties =
-    '{d7fc75d5-3492-5bbb-9b34-dac3e24e79d0}';
-const IID_IIterable_IMediaMarker = '{a1c0a397-0364-5e4c-9dca-7cd7011bd114}';
-const IID_IIterable_INT64 = '{7784427e-f9cc-518d-964b-e50d5ce727f1}';
-const IID_IIterable_IPlayReadyDomain = '{84e98f86-4bee-5f41-93a1-255887122d9f}';
-const IID_IIterable_IPlayReadyLicense =
-    '{9f28f6b7-b5ea-5073-ba3d-8cb2f07291a1}';
-const IID_IIterable_IPlayReadySecureStopServiceRequest =
-    '{8d8ac279-b07d-5308-9c78-1c5c996ca03c}';
-const IID_IIterable_IPropertySet = '{489b756d-be43-5abb-b9a0-a47254103339}';
-const IID_IIterable_IPwmControllerProvider =
-    '{4936ed59-b494-5128-bc7e-03e630346475}';
-const IID_IIterable_IRandomAccessStream =
-    '{ba666a00-1555-5df4-81a5-07d23f7ffceb}';
-const IID_IIterable_IRemoteSystemFilter =
-    '{13966c92-a8de-50c0-b16b-00c2c48f5f37}';
-const IID_IIterable_ISmsBinaryMessage =
-    '{5678a6a5-4d5a-51c2-a133-4b83bf25d987}';
-const IID_IIterable_ISmsMessage = '{ecabfd70-9601-5e38-83cf-b1046022a244}';
-const IID_IIterable_ISpeechRecognitionConstraint =
-    '{88e6436c-3253-520b-9ed8-a63b178c44a2}';
-const IID_IIterable_ISpiControllerProvider =
-    '{71ba027d-8c84-58b1-8d66-9177c11698eb}';
-const IID_IIterable_IStorageFile = '{76d43c7e-fd09-5908-a2b9-a49b4848294b}';
-const IID_IIterable_IStorageItem = '{bb8b8418-65d1-544b-b083-6d172f568c73}';
-const IID_IIterable_IStorageItemInformation =
-    '{43bc252e-a3d6-5f00-a12c-b088d3b912d4}';
-const IID_IIterable_ISyndicationNode = '{b486569a-72b3-57aa-9950-cea0b3e4fc58}';
-const IID_IIterable_IVideoEffectDefinition =
-    '{eb567f6f-b014-513d-99cd-f16c226c3c41}';
-const IID_IIterable_IVpnCustomPrompt = '{8bac695c-70cb-54d6-8606-af6e3a25e3a1}';
-const IID_IIterable_IVpnCustomPromptElement =
-    '{437d3693-00c4-50b4-989a-938f1016a230}';
-const IID_IIterable_IVpnProfile = '{db35f6b1-f266-5c87-8862-9dd87d9df18f}';
-const IID_IIterable_IWwwFormUrlDecoderEntry =
-    '{876be83b-7218-5bfb-a169-83152ef7e146}';
-const IID_IIterable_IXmlNode = '{f1146ffc-8c92-56e8-93f1-711f86722633}';
-const IID_IIterable_ImageStream = '{034ea0c4-c20e-5c0c-ba31-64212f28e650}';
-const IID_IIterable_IndexedResourceCandidate =
-    '{0fcb4184-1489-5774-9910-bab326bb50f6}';
-const IID_IIterable_IndexedResourceQualifier =
-    '{e16e0455-7341-55c0-a024-26d9ad43c8cc}';
-const IID_IIterable_InfraredTorchMode =
-    '{58a4b7b6-36c3-5541-b383-3690d7293c10}';
-const IID_IIterable_InstalledDesktopApp =
-    '{b64037f2-dc1a-57de-8b03-18a16f9ddbdf}';
-const IID_IIterable_IppAttributeValue =
-    '{2bd5bedc-2312-564d-b685-68684f631e92}';
-const IID_IIterable_IppIntegerRange = '{c1ab4df0-4e73-5bdd-80d1-e32e0264ce41}';
-const IID_IIterable_IppResolution = '{60cf8d08-474e-54fe-af66-6893e8f92575}';
-const IID_IIterable_IppTextWithLanguage =
-    '{da3d8cb6-082a-5993-8640-ac246fd71244}';
-const IID_IIterable_IsoSpeedPreset = '{94839abe-9712-545a-a94d-a567a3e8dfb7}';
-const IID_IIterable_IsolatedWindowsEnvironment =
-    '{37505de7-601e-5c17-a7f5-1a6ac4b91b03}';
-const IID_IIterable_IsolatedWindowsEnvironmentHostError =
-    '{e2c0cc79-8015-5196-8b32-7a0887a091bd}';
-const IID_IIterable_JapanesePhoneme = '{1aad17cb-1829-5236-8aef-0b75f8dfd7a6}';
-const IID_IIterable_LampArrayEffectPlaylist =
-    '{2314acda-c5df-5051-977d-94d79d1312fb}';
-const IID_IIterable_LanIdentifier = '{accef3cd-5d92-5c01-8ac4-79fe74cd733e}';
-const IID_IIterable_Language = '{48409a10-61b6-5db1-a69d-8abc46ac608a}';
-const IID_IIterable_LocalLocation = '{5474eefb-60f4-58ac-89e2-6e83f79fa76c}';
-const IID_IIterable_LocalLocationHoursOfOperationItem =
-    '{f298f515-b9bd-5297-b8bd-1c0040daac76}';
-const IID_IIterable_LockScreenBadge = '{6e82dedc-b74e-503a-b00b-9c6f47f12a0f}';
-const IID_IIterable_ManeuverWarning = '{ce0a7c13-d3c4-55af-a90f-c53f7bd93373}';
-const IID_IIterable_ManualFocusDistance =
-    '{cf8cbeb1-2a4c-522d-962f-84c31a598d68}';
-const IID_IIterable_MapLocation = '{77da6151-0763-508a-9041-3310baace575}';
-const IID_IIterable_MapRoute = '{d88a62a2-0edf-5312-97a8-10aeaea80b99}';
-const IID_IIterable_MapRouteLeg = '{8ff98759-78cd-56e8-877b-83ce846d6f8b}';
-const IID_IIterable_MapRouteManeuver = '{de9015fb-91d7-556e-bb4d-200b6f58fad4}';
-const IID_IIterable_MdmAlert = '{a0a617dc-210c-529f-b5e9-29aeceebb5a8}';
-const IID_IIterable_MediaBreak = '{04f000ec-c727-5dc0-9f7c-245c75c92f2f}';
-const IID_IIterable_MediaCaptureVideoProfile =
-    '{dbd49d71-e07b-5e11-824a-62ace8bdfc3e}';
-const IID_IIterable_MediaCaptureVideoProfileMediaDescription =
-    '{0895e56d-fe1f-5364-ab67-c597d8970b89}';
-const IID_IIterable_MediaClip = '{ff2e9b5c-26d0-575d-a3eb-7d938bd16f17}';
-const IID_IIterable_MediaFrameFormat = '{1f029a27-1123-538a-9261-8a380e12bac6}';
-const IID_IIterable_MediaFrameSource = '{0d447be3-8c91-581d-8071-17984b8b5994}';
-const IID_IIterable_MediaFrameSourceGroup =
-    '{d0b71deb-76e8-5833-9623-2b1e1a8e1b72}';
-const IID_IIterable_MediaFrameSourceInfo =
-    '{2e29c5b0-6aa9-50f2-91a4-5b67a5598f2e}';
-const IID_IIterable_MediaOverlay = '{efcda247-a1fb-51dc-a776-e3e06695fb36}';
-const IID_IIterable_MediaOverlayLayer =
-    '{7d2312d0-f3a2-5091-8a5e-41832e632c08}';
-const IID_IIterable_MediaPlaybackItem =
-    '{8b8ab4a4-0253-56e9-a232-94e335aefa8f}';
-const IID_IIterable_MediaTimeRange = '{268204e2-9d52-5e7b-a54b-869c63defc8e}';
-const IID_IIterable_MicrosoftAccountMultiFactorSessionInfo =
-    '{85a662c8-8a5d-59a8-9f73-ee237393c55c}';
-const IID_IIterable_MiracastReceiverConnection =
-    '{e989eb90-1f6f-5084-9bfb-1a5decca4f23}';
-const IID_IIterable_MiracastTransmitter =
-    '{5a3f4149-9b6b-5ff0-85a0-fea37b0990eb}';
-const IID_IIterable_MobileBroadbandAntennaSar =
-    '{fd66b9ac-40dc-5ac7-aaf1-2d3403e5fcbb}';
-const IID_IIterable_MobileBroadbandCellCdma =
-    '{46e83a22-4c40-5f27-bbcd-255dfd97ea93}';
-const IID_IIterable_MobileBroadbandCellGsm =
-    '{83e5eae8-3887-599e-bebf-8c51362db44c}';
-const IID_IIterable_MobileBroadbandCellLte =
-    '{45d961d3-e228-5afd-b18c-d4cfa3903432}';
-const IID_IIterable_MobileBroadbandCellNR =
-    '{f7c1a3a9-b0cd-5158-ab79-6c99f3cb7a4f}';
-const IID_IIterable_MobileBroadbandCellTdscdma =
-    '{6e1e543f-1cf0-5cb3-b3fc-b559213c58e2}';
-const IID_IIterable_MobileBroadbandCellUmts =
-    '{20392566-69cb-5eda-b641-5510e7ed1a12}';
-const IID_IIterable_MobileBroadbandDeviceServiceInformation =
-    '{88511855-6fe6-5694-83a7-991e29033de5}';
-const IID_IIterable_MobileBroadbandNetworkRegistrationStateChange =
-    '{0b90bb30-660c-51c6-9b8c-31dd8486e10e}';
-const IID_IIterable_MobileBroadbandPinLockStateChange =
-    '{aa4a8700-9943-59a3-8647-d373fd5e0e2b}';
-const IID_IIterable_MobileBroadbandPinType =
-    '{9d55726d-813e-50fb-9498-87aa872dd6ca}';
-const IID_IIterable_MobileBroadbandRadioStateChange =
-    '{c385adaa-574c-5ad8-98c2-61309525132d}';
-const IID_IIterable_MobileBroadbandSlotInfo =
-    '{30f9a178-40f1-5c82-865b-06660e680c35}';
-const IID_IIterable_MobileBroadbandUiccApp =
-    '{af538114-bd14-53b0-b1d1-841dcaa451ad}';
-const IID_IIterable_MseSourceBuffer = '{5ca0da86-6307-558a-865f-3a1ef9fe56b3}';
-const IID_IIterable_MseTimeRange = '{ed0c6e08-e22f-5ac2-ba14-e0ff1df3467f}';
-const IID_IIterable_NamedResource = '{e80d3d9d-96c9-579e-8e42-d550700de925}';
-const IID_IIterable_NetworkOperatorTetheringClient =
-    '{4762ecb3-af48-5b63-89b7-78a42056549f}';
-const IID_IIterable_NetworkUsage = '{dd2656b1-8360-5772-b272-c47f7f0fc7a6}';
-const IID_IIterable_NitRange = '{7fb7a783-ce2d-552d-bee3-bc1442db0409}';
-const IID_IIterable_Object = '{092b849b-60b1-52be-a44a-6fe8e933cbe4}';
-const IID_IIterable_OcrLine = '{6afa94a2-60d7-5dbe-942d-81aa3929c85e}';
-const IID_IIterable_OcrWord = '{a0ce663a-46d0-55e5-928e-251eb67a1e99}';
-const IID_IIterable_OfflineMapPackage =
-    '{7522287c-0af2-586c-bb3f-457c07984a6f}';
-const IID_IIterable_OnlineIdServiceTicket =
-    '{809d1314-97ab-5544-9891-ddcdfadd1dbb}';
-const IID_IIterable_OnlineIdServiceTicketRequest =
-    '{cb72d686-9516-520d-a274-fa4cd1762cb2}';
-const IID_IIterable_OpticalImageStabilizationMode =
-    '{323d7734-94c2-544d-a560-56560fe68819}';
-const IID_IIterable_Package = '{69ad6aa7-0c49-5f27-a5eb-ef4d59467b6d}';
-const IID_IIterable_PackageContentGroup =
-    '{d7dd1456-4805-5768-a25d-99641b096491}';
-const IID_IIterable_PackageUserInformation =
-    '{341348b9-52c8-5b57-9e91-f19f2a05b188}';
-const IID_IIterable_PackageVolume = '{a6199162-b163-56a1-9980-db0c3f4e9284}';
-const IID_IIterable_PasswordCredential =
-    '{0d224a66-bad5-5ad5-9ade-1e9f5a60fe73}';
-const IID_IIterable_PaymentDetailsModifier =
-    '{585d2b3d-cb34-58c4-81f4-1ea157996def}';
-const IID_IIterable_PaymentItem = '{b61b704c-e5fa-5524-8b95-7d03f5d36ae9}';
-const IID_IIterable_PaymentMethodData =
-    '{8c9ed87e-4add-58a8-ad9a-9dfa48ca250a}';
-const IID_IIterable_PaymentShippingOption =
-    '{de881c69-6642-54de-a8f7-d1a88b2404cf}';
-const IID_IIterable_PedometerReading = '{bbb61a5c-98c3-5718-88fe-5392a7451e2d}';
-const IID_IIterable_PeerInformation = '{917e1105-0647-5793-9d50-befe225f0f2f}';
-const IID_IIterable_PerceptionColorFrameSource =
-    '{db18069e-7b5a-54c3-a627-d56f9517fdf5}';
-const IID_IIterable_PerceptionCorrelation =
-    '{ca6bf87e-1745-5cd0-aee2-59736f5a206d}';
-const IID_IIterable_PerceptionDepthFrameSource =
-    '{8a07e1e8-5a02-585b-a26e-ad79beaa94cf}';
-const IID_IIterable_PerceptionInfraredFrameSource =
-    '{9309d0db-338d-5adf-8b3e-509bfdfccef3}';
-const IID_IIterable_PerceptionVideoProfile =
-    '{f6aea351-eb9b-564d-b10a-06673094acc8}';
-const IID_IIterable_PhoneCall = '{c43448ad-3660-5f7f-9270-cc340ebc32c2}';
-const IID_IIterable_PhoneCallHistoryEntry =
-    '{a3f93eea-c846-52c7-aa5a-3306707f6369}';
-const IID_IIterable_PhotoImportItem = '{82347483-3b75-5e95-bba4-abc0b8a320aa}';
-const IID_IIterable_PhotoImportOperation =
-    '{94f33a8f-115a-50cb-b59d-ab8483a84842}';
-const IID_IIterable_PhotoImportSidecar =
-    '{2b7f92ad-e596-5669-b622-fbfbc7040e89}';
-const IID_IIterable_PhotoImportSource =
-    '{40e01d62-b413-5b43-ab07-ab28b23fc886}';
-const IID_IIterable_PhotoImportStorageMedium =
-    '{3233cbfe-f9ee-560f-bd0f-e36abe6cda7f}';
-const IID_IIterable_PhotoImportVideoSegment =
-    '{94dd3b44-da03-5d79-bbfb-1beaf2ede482}';
-const IID_IIterable_PlatformDiagnosticTraceInfo =
-    '{ecb0c107-c97b-52fe-a5e6-a33e93493769}';
-const IID_IIterable_PlaySpeed = '{c4a17a40-8c62-5884-822b-502526970b0d}';
-const IID_IIterable_PlaybackMediaMarker =
-    '{deeccfe9-6c01-576f-95f2-93515ae40ebf}';
-const IID_IIterable_PnpObject = '{30b50092-36ee-53ff-9450-029004436c60}';
-const IID_IIterable_Point = '{c192280d-3a09-5423-9dc5-67b83ebde41d}';
-const IID_IIterable_PointerDevice = '{ad26662c-845b-5c6d-aeaa-406f48c21ae9}';
-const IID_IIterable_PointerDeviceUsage =
-    '{592d6618-eaab-5a79-a47a-c7fc0b749a4e}';
-const IID_IIterable_PosPrinterRotation =
-    '{1b1e4d8d-15f5-5802-9b23-8b75ce2a58c5}';
-const IID_IIterable_PrintPageRange = '{ed6ddf38-34cf-5a2f-b9ac-413dffce7f81}';
-const IID_IIterable_PrintTicketOption =
-    '{5c7e6676-9046-5b6a-9eb0-c6a954e8226b}';
-const IID_IIterable_Printing3DBaseMaterial =
-    '{9a6bd130-6f22-559c-b92c-14f9f8ddda47}';
-const IID_IIterable_Printing3DBaseMaterialGroup =
-    '{c08f8e70-f6ef-5469-806a-7cb601dddb67}';
-const IID_IIterable_Printing3DColorMaterial =
-    '{c77d4f28-7882-52b4-b3c9-7d58c8836573}';
-const IID_IIterable_Printing3DColorMaterialGroup =
-    '{1bf32a86-26ab-5750-b54c-3bda67867f8a}';
-const IID_IIterable_Printing3DComponent =
-    '{516556ca-f862-59f8-8241-e0f0c177dadd}';
-const IID_IIterable_Printing3DComponentWithMatrix =
-    '{8a213648-0b81-5e23-a48e-afe9f6691cc1}';
-const IID_IIterable_Printing3DCompositeMaterial =
-    '{a0af2623-1b11-53cf-975d-64959386cdd3}';
-const IID_IIterable_Printing3DCompositeMaterialGroup =
-    '{f2ffef61-c254-58c0-8206-b3b3096be9cb}';
-const IID_IIterable_Printing3DMesh = '{a8018fda-de4d-56fa-8609-fd2298bfb558}';
-const IID_IIterable_Printing3DModelTexture =
-    '{94790870-6041-5d04-8699-17417117bb85}';
-const IID_IIterable_Printing3DMultiplePropertyMaterial =
-    '{0dfc274e-ae4d-5bbb-93a8-7dc9f84ddac3}';
-const IID_IIterable_Printing3DMultiplePropertyMaterialGroup =
-    '{23f7518e-2439-5573-a683-efca0c61a8d6}';
-const IID_IIterable_Printing3DTexture2CoordMaterial =
-    '{28373276-483c-5bd0-99c7-01bfa04a57d4}';
-const IID_IIterable_Printing3DTexture2CoordMaterialGroup =
-    '{00017a1d-96bc-5c0e-b786-594fb4d077b6}';
-const IID_IIterable_Printing3DTextureResource =
-    '{54e3a71d-eae0-5199-9728-fac964850ebb}';
-const IID_IIterable_ProcessDiagnosticInfo =
-    '{97b73627-b296-5076-b8cd-6bd8a240e966}';
-const IID_IIterable_ProviderNetworkUsage =
-    '{f79bc7ba-01df-51ec-bfaf-fd883f698e07}';
-const IID_IIterable_ProximitySensorReading =
-    '{301ebccf-11ab-5e90-98ee-bd99c0e3bb76}';
-const IID_IIterable_PwmController = '{1403a6ab-73cb-5805-9bbc-a0dd39d476b0}';
-const IID_IIterable_RacingWheel = '{9a7c3830-9a87-5287-a1e2-8a2af29cf68c}';
-const IID_IIterable_Radio = '{e82500af-1f53-504e-b8be-dac4fbb69084}';
-const IID_IIterable_RawGameController =
-    '{8f2f08cc-f4f4-5539-9357-1f07334d381f}';
-const IID_IIterable_RcsEndUserMessageAction =
-    '{b6d21a55-471e-5a25-acdd-7784f7b4d7ba}';
-const IID_IIterable_RcsTransport = '{184c2264-65fd-5ad4-b22f-cb34e8eaec63}';
-const IID_IIterable_RectInt32 = '{d6222360-b82e-5eed-9eab-2e275b36e47e}';
-const IID_IIterable_RegionOfInterest = '{d73144c7-9d75-5dfb-8040-626202dcf454}';
-const IID_IIterable_RemoteSystemApp = '{4cfc1093-6de6-5d4d-8073-30e64b4dfa15}';
-const IID_IIterable_RemoteSystemSessionParticipant =
-    '{00189d10-16ec-5d1a-8369-4870c69e52b3}';
-const IID_IIterable_ResourceCandidate =
-    '{d89c9498-8869-57f8-a883-9c2dfeecb6c6}';
-const IID_IIterable_ResourceCollection =
-    '{d4372a2d-7ab0-5d8e-bd5c-6e9c0a67a8d8}';
-const IID_IIterable_ResourceMap = '{8bbe1154-19aa-53e7-9d6e-dc7d358580f4}';
-const IID_IIterable_ResourceQualifier =
-    '{2eee68e2-687c-5f7a-a14e-588c4d3089e1}';
-const IID_IIterable_RevocationAndRenewalItem =
-    '{07015868-578c-556b-8a35-4039a35d1d92}';
-const IID_IIterable_RfcommDeviceService =
-    '{3378e9a6-f6e2-50ea-bfee-b8109631feca}';
-const IID_IIterable_SearchSuggestion = '{df956160-f155-5816-b38b-a2ff0629ba0e}';
-const IID_IIterable_SecondaryAuthenticationFactorInfo =
-    '{43b7bbe4-f096-53dd-8c16-1faa4b468c86}';
-const IID_IIterable_SelectableWordSegment =
-    '{784b0cec-1348-5334-91e5-2e013294e211}';
-const IID_IIterable_ShareProvider = '{0903b218-5cad-53e6-9a21-6f4b31c4a409}';
-const IID_IIterable_SharedPackageContainer =
-    '{68a1862f-3ffa-58ec-89a9-5dfeb6d7433d}';
-const IID_IIterable_SharedPackageContainerMember =
-    '{250d357d-d93f-5603-8c84-3d325bb319ee}';
-const IID_IIterable_SignalDetectorResourceKind =
-    '{263e7b6e-5338-5871-9fdb-a27d9bc02e92}';
-const IID_IIterable_SimpleHapticsController =
-    '{b50da692-4a2b-5c8a-8e14-0439c0b1dba4}';
-const IID_IIterable_SimpleHapticsControllerFeedback =
-    '{8894a0df-33b0-57b0-aa1a-9255eee72dd5}';
-const IID_IIterable_Size = '{c9df55c3-4d41-5e90-ba76-e89ed564446b}';
-const IID_IIterable_SizeUInt32 = '{14f5a281-66ef-5299-aed1-96a1dc1e8416}';
-const IID_IIterable_SmartCard = '{a32c5202-d113-535f-880e-50f3e5121ef8}';
-const IID_IIterable_SmartCardAppletIdGroupRegistration =
-    '{4d153aad-915c-59c9-98b9-a97bf3a570ad}';
-const IID_IIterable_SmartCardAutomaticResponseApdu =
-    '{db52d376-027e-5270-a457-fb8b4ae8958c}';
-const IID_IIterable_SmartCardCryptogramAlgorithm =
-    '{8e0de0e9-0742-559e-9b1b-460268622c1f}';
-const IID_IIterable_SmartCardCryptogramMaterialCharacteristics =
-    '{92058349-b443-52d8-8e46-9cf3815dd15a}';
-const IID_IIterable_SmartCardCryptogramMaterialPackageCharacteristics =
-    '{ab1dac61-1bd9-54a6-b407-1dc4e5b1a197}';
-const IID_IIterable_SmartCardCryptogramMaterialPackageConfirmationResponseFormat =
-    '{c40c4451-4ebb-5635-9c7d-33c8c5d37a09}';
-const IID_IIterable_SmartCardCryptogramMaterialPackageFormat =
-    '{3e241acc-1745-57ce-9368-21ba2130c3c1}';
-const IID_IIterable_SmartCardCryptogramMaterialType =
-    '{2d379f84-389c-5809-a2c6-919b47caab88}';
-const IID_IIterable_SmartCardCryptogramPlacementStep =
-    '{234ab631-ed5f-51bc-8a8a-d5f3495f32de}';
-const IID_IIterable_SmartCardCryptogramStorageKeyCapabilities =
-    '{983619f1-45b9-5557-9800-eaa2bca6da57}';
-const IID_IIterable_SmartCardCryptogramStorageKeyCharacteristics =
-    '{16d5cc89-2f6a-5779-8ef8-2d5a20078150}';
-const IID_IIterable_SmsBroadcastType = '{12276b75-173e-514b-98f0-8a7927a9206c}';
-const IID_IIterable_SmsFilterRule = '{03ed8267-9c90-5260-8bc0-6c3e3306c964}';
-const IID_IIterable_SmsMessageRegistration =
-    '{f836fa0a-770d-5e8f-8664-01c43f959eea}';
-const IID_IIterable_SocialFeedItem = '{ad33d864-9569-5e2d-bd72-182a8ff50cf6}';
-const IID_IIterable_SocialItemThumbnail =
-    '{5d102c6d-92c3-59f3-b1dc-5986c56445a5}';
-const IID_IIterable_SoftwareBitmap = '{22d3a30f-0898-5e94-99a3-afa5951dfcd4}';
-const IID_IIterable_SortEntry = '{35aff6f9-ef75-5280-bb84-a2bf8317cf35}';
-const IID_IIterable_SpatialBoundingVolume =
-    '{89e8f1ee-3a2a-5b69-a786-cddcf7456a3a}';
-const IID_IIterable_SpeechRecognitionResult =
-    '{0d9b7b48-98a1-5b22-9a66-6f812f5947aa}';
-const IID_IIterable_SpiController = '{7b076938-dc1b-5368-9003-059291d37f35}';
-const IID_IIterable_StartupTask = '{61885ead-bf9e-5e9f-af04-6296b336930b}';
-const IID_IIterable_StorageFile = '{9ac00304-83ea-5688-87b6-ae38aab65d0b}';
-const IID_IIterable_StorageFolder = '{4669befc-ae5c-52b1-8a97-5466ce61e94e}';
-const IID_IIterable_StorageLibrary = '{851e3cfd-306b-5c8e-ae3c-a8d83c623604}';
-const IID_IIterable_StorageLibraryChange =
-    '{87c15dfc-0c5e-518b-9206-97d3d9823c61}';
-const IID_IIterable_StorageProviderError =
-    '{2b2a9a00-9929-5154-a897-1e6cfd7cf21a}';
-const IID_IIterable_StorageProviderFileTypeInfo =
-    '{b21025bc-4365-5728-b52e-51475166f141}';
-const IID_IIterable_StorageProviderItemProperty =
-    '{4584cb69-ee26-59e0-b05d-c9a7851a7317}';
-const IID_IIterable_StorageProviderItemPropertyDefinition =
-    '{41e78b90-1a7f-5dab-a123-7d5f5011dfeb}';
-const IID_IIterable_StorageProviderSyncRootInfo =
-    '{62ba69a0-f65c-502c-9782-b4bc25194d11}';
-const IID_IIterable_StoreAvailability =
-    '{c58de1a0-25de-578b-bb69-e0a26d67b203}';
-const IID_IIterable_StoreImage = '{b2da6de8-ad55-56ce-8754-2c96f4fe1c2e}';
-const IID_IIterable_StorePackageUpdate =
-    '{6b076c51-849e-5ec5-aed5-9b0585591902}';
-const IID_IIterable_StorePackageUpdateStatus =
-    '{60832223-7eb4-5cd7-8340-f5077173d364}';
-const IID_IIterable_StorePreviewProductInfo =
-    '{a9cb4860-67d1-53c2-a621-3074b0344d49}';
-const IID_IIterable_StorePreviewSkuInfo =
-    '{2c0d85d9-0df7-5de6-962e-bc8f149faf19}';
-const IID_IIterable_StoreQueueItem = '{68eb92e6-3cc6-5259-9a05-bd7f8d9fb8da}';
-const IID_IIterable_StoreSku = '{522d34ef-4b5a-5b44-a046-7a16051d011e}';
-const IID_IIterable_StoreSystemFeature =
-    '{98a8577a-b128-5400-8d3d-58654eaaf957}';
-const IID_IIterable_StoreVideo = '{46be61e4-7173-565e-aed5-4a2152f1ce69}';
-const IID_IIterable_String = '{e2fcc7c1-3bfc-5a0b-b2b0-72e769d1cb7e}';
-const IID_IIterable_SyndicationAttribute =
-    '{329eabe1-efcc-539e-96ba-f6a44f221dbd}';
-const IID_IIterable_SyndicationCategory =
-    '{d151f7d1-eabd-5300-b55c-149eb289cc71}';
-const IID_IIterable_SyndicationItem = '{55463eef-ecb8-59cd-8d6b-74daacbe7d19}';
-const IID_IIterable_SyndicationLink = '{c6919f6a-66d9-556a-9632-87d39af14638}';
-const IID_IIterable_SyndicationPerson =
-    '{e58e7844-eb34-5284-b09e-de6762d548ca}';
-const IID_IIterable_SystemUpdateItem = '{318a3078-918c-5521-b460-0b4210360aa1}';
-const IID_IIterable_TargetedContentAction =
-    '{cf05b497-3afd-5d00-859e-9fbd1a36d576}';
-const IID_IIterable_TargetedContentCollection =
-    '{2049f813-37ee-5158-9996-709859f0ce49}';
-const IID_IIterable_TargetedContentFile =
-    '{5f65d649-ccbd-5728-a85b-d3ff92fca962}';
-const IID_IIterable_TargetedContentImage =
-    '{efadb6bf-af18-5af9-a509-19881bc586f5}';
-const IID_IIterable_TargetedContentItem =
-    '{03f38fb6-54e6-5bf1-913b-9510fec8be1f}';
-const IID_IIterable_TextPhoneme = '{2017f561-8162-55f8-b056-656636935815}';
-const IID_IIterable_TextSegment = '{5498f4f3-cee4-5b72-9729-815c4ad7b9dc}';
-const IID_IIterable_TimeSpan = '{e9f78726-829a-5f67-8d19-95ef154b7742}';
-const IID_IIterable_TimedMetadataStreamDescriptor =
-    '{f3d07841-3852-509d-a12b-a9f2ac89da93}';
-const IID_IIterable_TimedMetadataTrack =
-    '{150a4454-03bb-5dd1-8153-6a600e851f71}';
-const IID_IIterable_TimedTextLine = '{aec710ad-3bd8-5a59-ae77-e7fc46fb105b}';
-const IID_IIterable_TimedTextSource = '{4fc57282-0edf-5d85-9d89-ddc2a5692c13}';
-const IID_IIterable_TimedTextSubformat =
-    '{cee3ae43-c93c-56f7-8c55-9024ded72c76}';
-const IID_IIterable_UINT16 = '{ecfa9a6f-fa2e-5345-b297-efb4e8c6be87}';
-const IID_IIterable_UINT32 = '{421d4b91-b13b-5f37-ae54-b5249bd80539}';
-const IID_IIterable_UINT64 = '{4b3a3229-7995-5f3c-b248-6c1f7e664f01}';
-const IID_IIterable_UINavigationController =
-    '{8dea85a0-0204-57dd-abad-90e37c0ef240}';
-const IID_IIterable_UnfulfilledConsumable =
-    '{2f4d1483-dd86-5fdb-8c44-06c98844bf3d}';
-const IID_IIterable_UnsupportedAppRequirement =
-    '{bd9ded1d-0379-5143-a490-168b6b8413a3}';
-const IID_IIterable_UploadOperation = '{79778799-38cc-5b67-9cd0-043fc47a9ef7}';
-const IID_IIterable_Uri = '{b0d63b78-78ad-5e31-b6d8-e32a0e16c447}';
-const IID_IIterable_UsbBulkInEndpointDescriptor =
-    '{101b1fd9-f1c9-5dda-9ad4-71176fa839b2}';
-const IID_IIterable_UsbBulkInPipe = '{2201a671-42d2-508d-a848-64b5447083c8}';
-const IID_IIterable_UsbBulkOutEndpointDescriptor =
-    '{b80beb39-62b3-5f59-b3e7-882cc9c5b0c0}';
-const IID_IIterable_UsbBulkOutPipe = '{9824caba-5ca6-5c2d-80cf-1949026d7857}';
-const IID_IIterable_UsbDescriptor = '{989909a5-5a03-51fb-bd94-84da7bda8819}';
-const IID_IIterable_UsbInterface = '{f54037ed-92e9-590d-b904-3ad7bfa9a621}';
-const IID_IIterable_UsbInterfaceSetting =
-    '{1aaf5739-9c2c-533e-a0e9-d53fdb45d15d}';
-const IID_IIterable_UsbInterruptInEndpointDescriptor =
-    '{8a7bac69-1f10-59c7-9837-72cfed7154a4}';
-const IID_IIterable_UsbInterruptInPipe =
-    '{39aef336-18aa-5be4-86d9-e332fe2632f3}';
-const IID_IIterable_UsbInterruptOutEndpointDescriptor =
-    '{09393d62-2316-536b-8a10-7038884ab2a7}';
-const IID_IIterable_UsbInterruptOutPipe =
-    '{e61a011e-4abe-53f2-83b3-ed4a949d2e3f}';
-const IID_IIterable_User = '{d1bacd1f-0376-5823-8c29-1d45b9f4c191}';
-const IID_IIterable_UserActivity = '{37e6ab55-1f30-5622-9778-1bdc07ac799f}';
-const IID_IIterable_UserActivitySessionHistoryItem =
-    '{29439e38-9cf0-51c7-a549-4469039caf79}';
-const IID_IIterable_UserDataAccount = '{1bc7f1ff-26b4-5440-9629-cb4cd8c87987}';
-const IID_IIterable_UserDataAccountPartnerAccountInfo =
-    '{5cdb425e-da5a-55fa-b349-5467996cab32}';
-const IID_IIterable_UserDataTask = '{8ae8becd-375f-5932-91ed-f82a74224365}';
-const IID_IIterable_UserDataTaskList = '{2c4d63bc-cffb-50d2-8a82-a9aefad651d3}';
-const IID_IIterable_UserWatcherUpdateKind =
-    '{89342b56-eb84-54e7-98c1-baee448bdced}';
-const IID_IIterable_VibrationDevice = '{1a40c994-8810-5688-9362-c4bb51018552}';
-const IID_IIterable_VideoEncodingProperties =
-    '{73c7317c-8682-5f81-84a2-30c425fa2d24}';
-const IID_IIterable_VideoStreamDescriptor =
-    '{3acbf03c-0a79-5823-aaa9-d88bc3f8f594}';
-const IID_IIterable_VideoTemporalDenoisingMode =
-    '{caf26629-ee84-5d4c-ae37-9dc4b26978eb}';
-const IID_IIterable_VideoTrack = '{84413442-a2e3-5e0f-936d-bc40b0fb2dcd}';
-const IID_IIterable_VoiceCommandContentTile =
-    '{bd13249b-8099-5573-bf74-7457796e92e5}';
-const IID_IIterable_VoiceInformation = '{3c33bb52-bd98-5c8c-adee-ee8da0628efc}';
-const IID_IIterable_VpnAppId = '{0e1e00aa-f93d-5dc7-9912-e07d1fa6bd67}';
-const IID_IIterable_VpnDomainNameInfo =
-    '{33abe488-be1a-558a-a9cf-b5330ab49f50}';
-const IID_IIterable_VpnNamespaceInfo = '{17781d03-ddcf-553f-aba8-d2e8155cb6b8}';
-const IID_IIterable_VpnPacketBuffer = '{fa954f6e-30d7-50e0-9d43-dadb6c53e196}';
-const IID_IIterable_VpnRoute = '{bbf498d5-b9ef-55f1-97b1-77a06639e4e2}';
-const IID_IIterable_VpnTrafficFilter = '{b6d4c853-77c9-52ca-9ce9-853add4554cf}';
-const IID_IIterable_WalletItem = '{ac174c8c-0fdd-5cff-a29f-4e8ce1c8bc81}';
-const IID_IIterable_WebAccount = '{cb15d439-a910-542a-89ed-7cfe67848a83}';
-const IID_IIterable_WebAccountClientView =
-    '{610e0f9d-aae4-54e1-bb0b-1aca14110abf}';
-const IID_IIterable_WebErrorStatus = '{7b7f182e-a6ce-556b-9a2e-ef97662f2aee}';
-const IID_IIterable_WebProviderTokenResponse =
-    '{e9bac236-c077-553a-b4ae-b58fb0b89918}';
-const IID_IIterable_WebTokenResponse = '{7e5bb7ec-bbd7-5575-9a61-f5815fa22a0e}';
-const IID_IIterable_WebViewControl = '{f1028969-d7aa-59d8-8826-4e02a8064515}';
-const IID_IIterable_WebViewControlDeferredPermissionRequest =
-    '{91805f3f-13cb-5483-b592-b9ae0e334f17}';
-const IID_IIterable_WiFiAdapter = '{e0bc76c4-8d0c-53fc-bcd4-228f47210ace}';
-const IID_IIterable_WiFiAvailableNetwork =
-    '{f17484ea-c71e-5d3e-b74c-3a0e61dd9c20}';
-const IID_IIterable_WiFiDirectConfigurationMethod =
-    '{794f12da-2dc6-5277-82dc-b0781610537b}';
-const IID_IIterable_WiFiDirectInformationElement =
-    '{19c1ca4e-9561-5253-96d9-dbaf28d47d89}';
-const IID_IIterable_WiFiDirectServiceConfigurationMethod =
-    '{d9773b1a-a148-58bf-9c4b-afeac9be3ab4}';
-const IID_IIterable_WiFiWpsKind = '{41e16513-a8f2-55ed-9be4-5665167d49d7}';
-const IID_IIterable_WordSegment = '{d14dc94a-a311-5ddc-bf8b-d58795d87d6d}';
-const IID_IIterable_Workspace = '{f02d0ebe-eac2-502f-9836-1c5482333bfe}';
-const IID_IIterable_XboxLiveDeviceAddress =
-    '{44df817c-b475-5fdb-862a-4a96ed1dad5c}';
-const IID_IIterable_XboxLiveEndpointPair =
-    '{7fcaf666-ddc0-50dc-a76f-9bb6058be3c9}';
-const IID_IIterable_XboxLiveEndpointPairTemplate =
-    '{9fce55c3-5580-581a-85b5-88dd5e9d4ff7}';
-const IID_IIterable_XboxLiveQualityOfServiceMetric =
-    '{e202eef6-93bd-5fa2-91fc-7ca6a4edc38a}';
-const IID_IIterable_XboxLiveQualityOfServiceMetricResult =
-    '{65cc99ec-9779-568a-8982-aee439fa24ce}';
-const IID_IIterable_XboxLiveQualityOfServicePrivatePayloadResult =
-    '{3a061dd4-03dd-52fb-9a83-067de27c2bb1}';
-const IID_IIterable_ZoomTransitionMode =
-    '{db656915-8fac-5fb2-98e0-0e97421656c5}';
-const IID_IIterable_bool = '{30160817-1d7d-54e9-99db-d7636266a476}';
-const IID_IIterable_byte = '{88318266-f3fd-50fc-8f08-b823a41b60c1}';
-const IID_IIterable_double = '{c738964e-9c64-5bce-b5ce-61e9a282ec4a}';
-const IID_IIterable_float = '{b01bee51-063a-5fda-bd72-d76637bb8cb8}';
-const IID_IIterable_int = '{81a643fb-f51c-5565-83c4-f96425777b66}';
-const IID_IIterable_short = '{72ff2923-4b4e-53bb-8feb-41ec5f2bb734}';
-
-// IKeyValuePair<K, V> IIDs
-const IID_IKeyValuePair = '{02b51929-c1c4-4a7e-8940-0312b5c18500}';
-
-// IMap<K, V> IIDs
-const IID_IMap = '{3c2925fe-8519-45c1-aa79-197b6718c1c1}';
-const IID_IMap_Guid_Object = '{5ee3189c-7dbf-5998-ad07-5414fb82567c}';
-const IID_IMap_Object_Object = '{f5f69427-55ed-5512-8429-d4f6626dfcdd}';
-const IID_IMap_String_String = '{f6d1f700-49c2-52ae-8154-826f9908773c}';
-const IID_IMap_String_Object = '{1b0d3570-0877-5ec2-8a2c-3b9539506aca}';
-const IID_IMap_String_IJsonValue = '{c9d9a725-786b-5113-b4b7-9b61764c220b}';
-
-// IMapView<K, V> IIDs
-const IID_IMapView = '{e480ce40-a338-4ada-adcf-272272e48cb9}';
-
-// IObservableMap<K, V> IIDs
-const IID_IObservableMap_String_String =
-    '{1e036276-2f60-55f6-b7f3-f86079e6900b}';
-const IID_IObservableMap_String_Object =
-    '{236aac9d-fb12-5c4d-a41c-9e445fb4d7ec}';
-
-// IObservableVector<T> IIDs
-const IID_IObservableVector_SearchSuggestion =
-    '{e885956c-72c1-514f-ad9a-2943b880ae13}';
-const IID_IObservableVector_TimedMetadataTrack =
-    '{98ebeba8-6659-5af2-be4f-32fdeede3b77}';
-const IID_IObservableVector_TimedTextSource =
-    '{95e6c852-6e9c-50a9-8935-054320242945}';
-const IID_IObservableVector_MediaPlaybackItem =
-    '{fabf4843-39eb-5c3d-a12b-fca189b46e8e}';
-const IID_IObservableVector_StorageFolder =
-    '{358f44df-2a45-5cb8-9385-1ff66808cde0}';
-const IID_IObservableVector_Object = '{7b81c56a-0985-518d-baa9-0da9ae009f65}';
-
-// IReference<T> IIDs
-const IID_IReference = '{61c17706-2d65-11e0-9ae8-d48564015472}';
-const IID_IReference_AdaptiveMediaSourceResourceType =
-    '{74c8c3aa-de03-5bf0-aae8-aa8b692066b3}';
-const IID_IReference_BasicGeoposition =
-    '{e4d5dda6-f57c-57cc-b67f-2939a901dabe}';
-const IID_IReference_BitmapBounds = '{aa19da70-dee6-5b42-b562-2fcd218c34ca}';
-const IID_IReference_BluetoothLEAdvertisementFlags =
-    '{91c0ba96-9e69-5b82-bf1d-83ab2a509c53}';
-const IID_IReference_Boolean = '{3c00fd60-2950-5939-a21a-2d12c5a01b8a}';
-const IID_IReference_CaptureSceneMode =
-    '{e20596aa-0bbe-5203-be6b-6b71ff5b0843}';
-const IID_IReference_DateTime = '{5541d8a7-497c-5aa4-86fc-7713adbf2a2c}';
-const IID_IReference_DisplayPresentationRate =
-    '{08882ddb-a981-5887-b751-106993d66229}';
-const IID_IReference_Double = '{2f2d6c29-5473-5f3e-92e7-96572bb990e2}';
-const IID_IReference_EmailMailboxSmimeEncryptionAlgorithm =
-    '{433b124d-60e6-5b7a-96bc-8681de5f5e75}';
-const IID_IReference_EmailMailboxSmimeSigningAlgorithm =
-    '{e498a622-f168-5104-a53b-faf996f51632}';
-const IID_IReference_Float = '{719cc2ba-3e76-5def-9f1a-38d85a145ea8}';
-const IID_IReference_Guid = '{7d50f649-632c-51f9-849a-ee49428933ea}';
-const IID_IReference_HdcpProtection = '{8e330979-2fef-5d68-88aa-a9ee6697d117}';
-const IID_IReference_HolographicStereoTransform =
-    '{6e67ce78-cc67-52c0-b635-991db0bff5ca}';
-const IID_IReference_Int16 = '{6ec9e41b-6709-5647-9918-a1270110fc4e}';
-const IID_IReference_Int32 = '{548cefbd-bc8a-5fa0-8df2-957440fc8bf4}';
-const IID_IReference_Int64 = '{4dda9e24-e69f-5c6a-a0a6-93427365af2a}';
-const IID_IReference_ManualFocusDistance =
-    '{b0060b8a-1105-5ad4-963d-f6cf1905d349}';
-const IID_IReference_Matrix4x4 = '{dacbffdc-68ef-5fd0-b657-782d0ac9807e}';
-const IID_IReference_MediaCaptureFocusState =
-    '{58820185-5da0-5faa-86da-9bd9f03974fa}';
-const IID_IReference_MediaPlaybackAutoRepeatMode =
-    '{50a7f41f-58d5-5c4d-9475-8dd1acd65836}';
-const IID_IReference_MediaPlaybackType =
-    '{e289f7d8-6ba7-50ab-9f13-6e4e51d15ca4}';
-const IID_IReference_MseTimeRange = '{2fffd101-16f8-596d-a88e-659b6f584641}';
-const IID_IReference_Point = '{84f14c22-a00a-5272-8d3d-82112e66df00}';
-const IID_IReference_Quaternion = '{b27004bb-c014-5dce-9a21-799c5a3c1461}';
-const IID_IReference_Rect = '{80423f11-054f-5eac-afd3-63b6ce15e77b}';
-const IID_IReference_Size = '{61723086-8e53-5276-9f36-2a4bb93e2b75}';
-const IID_IReference_SizeInt32 = '{b77aa86d-2d6e-55f1-8f99-64ac5c05328b}';
-const IID_IReference_SpatialBoundingBox =
-    '{ab3274d9-9b82-5396-bb00-d70c539796b3}';
-const IID_IReference_SpatialBoundingFrustum =
-    '{f434face-0c36-5749-a8a0-0bb6ce78a614}';
-const IID_IReference_SpatialBoundingOrientedBox =
-    '{09f88309-9f81-5207-bdb2-abef926db18f}';
-const IID_IReference_SpatialRay = '{44dd686b-c7d8-582c-91c2-d98e604dcfd4}';
-const IID_IReference_TimeSpan = '{604d0c4c-91de-5c2a-935f-362f13eaf800}';
-const IID_IReference_Uint8 = '{e5198cc8-2873-55f5-b0a1-84ff9e4aad62}';
-const IID_IReference_Uint32 = '{513ef3af-e784-5325-a91e-97c2b8111cf3}';
-const IID_IReference_Uint64 = '{6755e376-53bb-568b-a11d-17239868309e}';
-const IID_IReference_UserDataTaskDaysOfWeek =
-    '{903861c5-d29d-5083-a03d-6b2a0233dcca}';
-const IID_IReference_UserDataTaskWeekOfMonth =
-    '{5246fa7f-f9e9-5574-80f7-4aa359b0023d}';
-const IID_IReference_Vector2 = '{48f6a69e-8465-57ae-9400-9764087f65ad}';
-const IID_IReference_Vector3 = '{1ee770ff-c954-59ca-a754-6199a9be282c}';
-const IID_IReference_WebErrorStatus = '{f2b26336-6a9d-54de-8eca-00d6c871e469}';
-const IID_IReference_WhiteBalanceGain =
-    '{79c7838a-39e2-5287-ac3b-b118db251002}';
-
-// IVector<T> IIDs
-const IID_IVector = '{913337e9-11a1-4345-a3a2-4e7f956e222d}';
-const IID_IVector_IJsonValue = '{d44662bc-dce3-59a8-9272-4b210f33908b}';
-
-// IVectorView<T> IIDs
-const IID_IVectorView = '{bbe1fa4c-b0e3-4583-baef-1f1b2e483e56}';
-const IID_IVectorView_IWwwFormUrlDecoderEntry =
-    '{b1f00d3b-1f06-5117-93ea-2a0d79116701}';
