+++ packages/windows_foundation/lib/src/internal/iids.dart
@@ -1,0 +1,87 @@
+// Copyright (c) 2023, the dartwinrt authors. Please see the AUTHORS file for
+// details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// ignore_for_file: constant_identifier_names
+
+// IIterable<IKeyValuePair<K, V>> IIDs
+
+const IID_IIterable_IKeyValuePair_Guid_Object =
+    '{f3b20528-e3b3-5331-b2d0-0c2623aee785}';
+const IID_IIterable_IKeyValuePair_String_Object =
+    '{fe2f3d47-5d47-5499-8374-430c7cda0204}';
+const IID_IIterable_IKeyValuePair_String_String =
+    '{e9bdaaf0-cbf6-5c72-be90-29cbf3a1319b}';
+
+// IIterable<T> IIDs
+const IID_IIterable_HostName = '{9e5f3ed0-cf1c-5d38-832c-acea6164bf5c}';
+const IID_IIterable_String = '{e2fcc7c1-3bfc-5a0b-b2b0-72e769d1cb7e}';
+const IID_IIterable_Uri = '{b0d63b78-78ad-5e31-b6d8-e32a0e16c447}';
+
+// IMap<K, V> IIDs
+const IID_IMap_Guid_Object = '{5ee3189c-7dbf-5998-ad07-5414fb82567c}';
+const IID_IMap_String_String = '{f6d1f700-49c2-52ae-8154-826f9908773c}';
+const IID_IMap_String_Object = '{1b0d3570-0877-5ec2-8a2c-3b9539506aca}';
+
+// IReference<T> IIDs
+const IID_IReference_AdaptiveMediaSourceResourceType =
+    '{74c8c3aa-de03-5bf0-aae8-aa8b692066b3}';
+const IID_IReference_BasicGeoposition =
+    '{e4d5dda6-f57c-57cc-b67f-2939a901dabe}';
+const IID_IReference_BitmapBounds = '{aa19da70-dee6-5b42-b562-2fcd218c34ca}';
+const IID_IReference_BluetoothLEAdvertisementFlags =
+    '{91c0ba96-9e69-5b82-bf1d-83ab2a509c53}';
+const IID_IReference_Boolean = '{3c00fd60-2950-5939-a21a-2d12c5a01b8a}';
+const IID_IReference_CaptureSceneMode =
+    '{e20596aa-0bbe-5203-be6b-6b71ff5b0843}';
+const IID_IReference_DateTime = '{5541d8a7-497c-5aa4-86fc-7713adbf2a2c}';
+const IID_IReference_DisplayPresentationRate =
+    '{08882ddb-a981-5887-b751-106993d66229}';
+const IID_IReference_Double = '{2f2d6c29-5473-5f3e-92e7-96572bb990e2}';
+const IID_IReference_EmailMailboxSmimeEncryptionAlgorithm =
+    '{433b124d-60e6-5b7a-96bc-8681de5f5e75}';
+const IID_IReference_EmailMailboxSmimeSigningAlgorithm =
+    '{e498a622-f168-5104-a53b-faf996f51632}';
+const IID_IReference_Float = '{719cc2ba-3e76-5def-9f1a-38d85a145ea8}';
+const IID_IReference_Guid = '{7d50f649-632c-51f9-849a-ee49428933ea}';
+const IID_IReference_HdcpProtection = '{8e330979-2fef-5d68-88aa-a9ee6697d117}';
+const IID_IReference_HolographicStereoTransform =
+    '{6e67ce78-cc67-52c0-b635-991db0bff5ca}';
+const IID_IReference_Int16 = '{6ec9e41b-6709-5647-9918-a1270110fc4e}';
+const IID_IReference_Int32 = '{548cefbd-bc8a-5fa0-8df2-957440fc8bf4}';
+const IID_IReference_Int64 = '{4dda9e24-e69f-5c6a-a0a6-93427365af2a}';
+const IID_IReference_ManualFocusDistance =
+    '{b0060b8a-1105-5ad4-963d-f6cf1905d349}';
+const IID_IReference_Matrix4x4 = '{dacbffdc-68ef-5fd0-b657-782d0ac9807e}';
+const IID_IReference_MediaCaptureFocusState =
+    '{58820185-5da0-5faa-86da-9bd9f03974fa}';
+const IID_IReference_MediaPlaybackAutoRepeatMode =
+    '{50a7f41f-58d5-5c4d-9475-8dd1acd65836}';
+const IID_IReference_MediaPlaybackType =
+    '{e289f7d8-6ba7-50ab-9f13-6e4e51d15ca4}';
+const IID_IReference_MseTimeRange = '{2fffd101-16f8-596d-a88e-659b6f584641}';
+const IID_IReference_Point = '{84f14c22-a00a-5272-8d3d-82112e66df00}';
+const IID_IReference_Quaternion = '{b27004bb-c014-5dce-9a21-799c5a3c1461}';
+const IID_IReference_Rect = '{80423f11-054f-5eac-afd3-63b6ce15e77b}';
+const IID_IReference_Size = '{61723086-8e53-5276-9f36-2a4bb93e2b75}';
+const IID_IReference_SizeInt32 = '{b77aa86d-2d6e-55f1-8f99-64ac5c05328b}';
+const IID_IReference_SpatialBoundingBox =
+    '{ab3274d9-9b82-5396-bb00-d70c539796b3}';
+const IID_IReference_SpatialBoundingFrustum =
+    '{f434face-0c36-5749-a8a0-0bb6ce78a614}';
+const IID_IReference_SpatialBoundingOrientedBox =
+    '{09f88309-9f81-5207-bdb2-abef926db18f}';
+const IID_IReference_SpatialRay = '{44dd686b-c7d8-582c-91c2-d98e604dcfd4}';
+const IID_IReference_TimeSpan = '{604d0c4c-91de-5c2a-935f-362f13eaf800}';
+const IID_IReference_Uint8 = '{e5198cc8-2873-55f5-b0a1-84ff9e4aad62}';
+const IID_IReference_Uint32 = '{513ef3af-e784-5325-a91e-97c2b8111cf3}';
+const IID_IReference_Uint64 = '{6755e376-53bb-568b-a11d-17239868309e}';
+const IID_IReference_UserDataTaskDaysOfWeek =
+    '{903861c5-d29d-5083-a03d-6b2a0233dcca}';
+const IID_IReference_UserDataTaskWeekOfMonth =
+    '{5246fa7f-f9e9-5574-80f7-4aa359b0023d}';
+const IID_IReference_Vector2 = '{48f6a69e-8465-57ae-9400-9764087f65ad}';
+const IID_IReference_Vector3 = '{1ee770ff-c954-59ca-a754-6199a9be282c}';
+const IID_IReference_WebErrorStatus = '{f2b26336-6a9d-54de-8eca-00d6c871e469}';
+const IID_IReference_WhiteBalanceGain =
+    '{79c7838a-39e2-5287-ac3b-b118db251002}';
