+++ packages/windows_foundation/lib/src/internal/ipropertyvalue_helpers.dart
@@ -7,7 +7,6 @@ import 'dart:ffi';
 import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
 
-import '../constants.dart';
 import '../enums.g.dart';
 import '../helpers.dart';
 import '../iinspectable.dart';
@@ -17,6 +16,7 @@ import '../propertyvalue.dart';
 import '../structs.g.dart';
 import 'comobject_pointer.dart';
 import 'hstring_array.dart';
+import 'iids.dart';
 import 'int_array.dart';
 
 extension IPropertyValueHelper on IPropertyValue {
