+++ packages/winrtgen/lib/src/constants/iids.dart
@@ -8,4 +8,8 @@
 
 const IID_IIterable = '{faa585ea-6214-4217-afda-7f46de5869b3}';
 const IID_IKeyValuePair = '{02b51929-c1c4-4a7e-8940-0312b5c18500}';
+const IID_IMap = '{3c2925fe-8519-45c1-aa79-197b6718c1c1}';
+const IID_IMapView = '{e480ce40-a338-4ada-adcf-272272e48cb9}';
 const IID_IReference = '{61c17706-2d65-11e0-9ae8-d48564015472}';
+const IID_IVector = '{913337e9-11a1-4345-a3a2-4e7f956e222d}';
+const IID_IVectorView = '{bbe1fa4c-b0e3-4583-baef-1f1b2e483e56}';
